// frontend/src/api_gen/models/notification-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { TripNotificationDTO } from './trip-notification-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { TripStatus } from './trip-status';

/**
 * 
 * @export
 * @interface NotificationDTO
 */
export interface NotificationDTO {
    /**
     * 
     * @type {TripNotificationDTO}
     * @memberof NotificationDTO
     */
    'trip': TripNotificationDTO;
    /**
     * 
     * @type {string}
     * @memberof NotificationDTO
     */
    'previousTripStatusStatus': NotificationDTOPreviousTripStatusStatusEnum;
    /**
     * 
     * @type {TripStatus}
     * @memberof NotificationDTO
     */
    'nextTripStatus': TripStatus;
}

export const NotificationDTOPreviousTripStatusStatusEnum = {
    PendingDepartmentApproval: 'PENDING_DEPARTMENT_APPROVAL',
    DepartmentApprovalRejected: 'DEPARTMENT_APPROVAL_REJECTED',
    TravelApproved: 'TRAVEL_APPROVED',
    PendingExpenseApproval: 'PENDING_EXPENSE_APPROVAL',
    ExpenseApprovalRejected: 'EXPENSE_APPROVAL_REJECTED',
    PendingDirectorApproval: 'PENDING_DIRECTOR_APPROVAL',
    DirectorApprovalRejected: 'DIRECTOR_APPROVAL_REJECTED',
    AwaitingPayment: 'AWAITING_PAYMENT',
    Paid: 'PAID'
} as const;

export type NotificationDTOPreviousTripStatusStatusEnum = typeof NotificationDTOPreviousTripStatusStatusEnum[keyof typeof NotificationDTOPreviousTripStatusStatusEnum];